requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dnspython==2.4.2
idna==3.6
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
sodapy==2.2.0
urllib3==2.1.0
